build: spin off charm builder as a separate tool
Move charm building routiness to a stand-alone charm builder tool
invoked from the "build" command.
