+++ charmcraft/commands/build.py
@@ -16,16 +16,15 @@
 
 """Infrastructure for the 'build' command."""
 
-import errno
 import logging
 import os
 import pathlib
-import shutil
 import subprocess
+import sys
 import zipfile
 from typing import List, Optional
 
-from charmcraft import linters
+from charmcraft import charm_builder, linters
 from charmcraft.bases import check_if_base_matches_host
 from charmcraft.cmdbase import BaseCommand, CommandError
 from charmcraft.config import Base, BasesConfiguration, Config
@@ -35,7 +34,6 @@ from charmcraft.env import (
     get_managed_environment_project_path,
     is_charmcraft_running_in_managed_mode,
 )
-from charmcraft.jujuignore import JujuIgnore, default_juju_ignore
 from charmcraft.logsetup import message_handler
 from charmcraft.manifest import create_manifest
 from charmcraft.metadata import parse_metadata_yaml
@@ -45,7 +43,6 @@ from charmcraft.providers import (
     is_base_providable,
     launched_environment,
 )
-from charmcraft.utils import make_executable
 
 logger = logging.getLogger(__name__)
 
@@ -97,20 +94,6 @@ def format_charm_file_name(
     return "_".join([charm_name, _format_bases_config(bases_config)]) + ".charm"
 
 
-def _pip_needs_system():
-    """Determine whether pip3 defaults to --user, needing --system to turn it off."""
-    cmd = [
-        "python3",
-        "-c",
-        (
-            "from pip.commands.install import InstallCommand; "
-            'assert InstallCommand().cmd_opts.get_option("--system") is not None'
-        ),
-    ]
-    proc = subprocess.run(cmd, stdout=subprocess.DEVNULL, stderr=subprocess.DEVNULL)
-    return proc.returncode == 0
-
-
 def polite_exec(cmd):
     """Execute a command, only showing output if error."""
     logger.debug("Running external command %s", cmd)
@@ -148,7 +131,6 @@ class Builder:
         self.requirement_paths = args["requirement"]
 
         self.buildpath = self.charmdir / BUILD_DIRNAME
-        self.ignore_rules = self._load_juju_ignore()
         self.config = config
         self.metadata = parse_metadata_yaml(self.charmdir)
 
@@ -161,13 +143,44 @@ class Builder:
         """
         logger.debug("Building charm in %r", str(self.buildpath))
 
-        if self.buildpath.exists():
-            shutil.rmtree(str(self.buildpath))
-        self.buildpath.mkdir()
+        build_env = dict(LANG="C.UTF-8", LC_ALL="C.UTF-8")
+        for key in [
+            "PATH",
+            "SNAP",
+            "SNAP_ARCH",
+            "SNAP_NAME",
+            "SNAP_VERSION",
+            "http_proxy",
+            "https_proxy",
+        ]:
+            if key in os.environ:
+                build_env[key] = os.environ[key]
 
-        linked_entrypoint = self.handle_generic_paths()
-        self.handle_dispatcher(linked_entrypoint)
-        self.handle_dependencies()
+        env_flags = [f"{key}={value}" for key, value in build_env.items()]
+
+        # invoke the charm builder
+        build_cmd = [
+            "env",
+            "-i",
+            *env_flags,
+            sys.executable,
+            "-I",
+            charm_builder.__file__,
+            "--charmdir",
+            str(self.charmdir),
+            "--builddir",
+            str(self.buildpath),
+        ]
+
+        if self.entrypoint:
+            build_cmd.extend(["--entrypoint", str(self.entrypoint)])
+
+        for req in self.requirement_paths:
+            build_cmd.extend(["-r", str(req)])
+
+        retcode = polite_exec(build_cmd)
+        if retcode:
+            raise CommandError("problems running charm builder")
 
         linting_results = []
         # run linters, present them to the user according to their type, and fail if necessary
@@ -330,167 +343,6 @@ class Builder:
                     ) from error
 
         return charm_name
-
-    def _load_juju_ignore(self):
-        ignore = JujuIgnore(default_juju_ignore)
-        path = self.charmdir / ".jujuignore"
-        if path.exists():
-            with path.open("r", encoding="utf-8") as ignores:
-                ignore.extend_patterns(ignores)
-        return ignore
-
-    def create_symlink(self, src_path, dest_path):
-        """Create a symlink in dest_path pointing relatively like src_path.
-
-        It also verifies that the linked dir or file is inside the project.
-        """
-        resolved_path = src_path.resolve()
-        if self.charmdir in resolved_path.parents:
-            relative_link = relativise(src_path, resolved_path)
-            dest_path.symlink_to(relative_link)
-        else:
-            rel_path = src_path.relative_to(self.charmdir)
-            logger.warning(
-                "Ignoring symlink because targets outside the project: %r",
-                str(rel_path),
-            )
-
-    def handle_generic_paths(self):
-        """Handle all files and dirs except what's ignored and what will be handled later.
-
-        Works differently for the different file types:
-        - regular files: hard links
-        - directories: created
-        - symlinks: respected if are internal to the project
-        - other types (blocks, mount points, etc): ignored
-        """
-        logger.debug("Linking in generic paths")
-
-        for basedir, dirnames, filenames in os.walk(
-            str(self.charmdir), followlinks=False
-        ):
-            abs_basedir = pathlib.Path(basedir)
-            rel_basedir = abs_basedir.relative_to(self.charmdir)
-
-            # process the directories
-            ignored = []
-            for pos, name in enumerate(dirnames):
-                rel_path = rel_basedir / name
-                abs_path = abs_basedir / name
-
-                if self.ignore_rules.match(str(rel_path), is_dir=True):
-                    logger.debug(
-                        "Ignoring directory because of rules: %r", str(rel_path)
-                    )
-                    ignored.append(pos)
-                elif abs_path.is_symlink():
-                    dest_path = self.buildpath / rel_path
-                    self.create_symlink(abs_path, dest_path)
-                else:
-                    dest_path = self.buildpath / rel_path
-                    dest_path.mkdir(mode=abs_path.stat().st_mode)
-
-            # in the future don't go inside ignored directories
-            for pos in reversed(ignored):
-                del dirnames[pos]
-
-            # process the files
-            for name in filenames:
-                rel_path = rel_basedir / name
-                abs_path = abs_basedir / name
-
-                if self.ignore_rules.match(str(rel_path), is_dir=False):
-                    logger.debug("Ignoring file because of rules: %r", str(rel_path))
-                elif abs_path.is_symlink():
-                    dest_path = self.buildpath / rel_path
-                    self.create_symlink(abs_path, dest_path)
-                elif abs_path.is_file():
-                    dest_path = self.buildpath / rel_path
-                    try:
-                        os.link(str(abs_path), str(dest_path))
-                    except PermissionError:
-                        # when not allowed to create hard links
-                        shutil.copy2(str(abs_path), str(dest_path))
-                    except OSError as e:
-                        if e.errno != errno.EXDEV:
-                            raise
-                        shutil.copy2(str(abs_path), str(dest_path))
-                else:
-                    logger.debug("Ignoring file because of type: %r", str(rel_path))
-
-        # the linked entrypoint is calculated here because it's when it's really in the build dir
-        linked_entrypoint = self.buildpath / self.entrypoint.relative_to(self.charmdir)
-        return linked_entrypoint
-
-    def handle_dispatcher(self, linked_entrypoint):
-        """Handle modern and classic dispatch mechanisms."""
-        # dispatch mechanism, create one if wasn't provided by the project
-        dispatch_path = self.buildpath / DISPATCH_FILENAME
-        if not dispatch_path.exists():
-            logger.debug("Creating the dispatch mechanism")
-            dispatch_content = DISPATCH_CONTENT.format(
-                entrypoint_relative_path=linked_entrypoint.relative_to(self.buildpath)
-            )
-            with dispatch_path.open("wt", encoding="utf8") as fh:
-                fh.write(dispatch_content)
-                make_executable(fh)
-
-        # bunch of symlinks, to support old juju: verify that any of the already included hooks
-        # in the directory is not linking directly to the entrypoint, and also check all the
-        # mandatory ones are present
-        dest_hookpath = self.buildpath / HOOKS_DIR
-        if not dest_hookpath.exists():
-            dest_hookpath.mkdir()
-
-        # get those built hooks that we need to replace because they are pointing to the
-        # entrypoint directly and we need to fix the environment in the middle
-        current_hooks_to_replace = []
-        for node in dest_hookpath.iterdir():
-            if node.resolve() == linked_entrypoint:
-                current_hooks_to_replace.append(node)
-                node.unlink()
-                logger.debug(
-                    "Replacing existing hook %r as it's a symlink to the entrypoint",
-                    node.name,
-                )
-
-        # include the mandatory ones and those we need to replace
-        hooknames = MANDATORY_HOOK_NAMES | {x.name for x in current_hooks_to_replace}
-        for hookname in hooknames:
-            logger.debug("Creating the %r hook script pointing to dispatch", hookname)
-            dest_hook = dest_hookpath / hookname
-            if not dest_hook.exists():
-                relative_link = relativise(dest_hook, dispatch_path)
-                dest_hook.symlink_to(relative_link)
-
-    def handle_dependencies(self):
-        """Handle from-directory and virtualenv dependencies."""
-        logger.debug("Installing dependencies")
-
-        # virtualenv with other dependencies (if any)
-        if self.requirement_paths:
-            retcode = polite_exec(["pip3", "list"])
-            if retcode:
-                raise CommandError("problems using pip")
-
-            venvpath = self.buildpath / VENV_DIRNAME
-            cmd = [
-                "pip3",
-                "install",  # base command
-                "--target={}".format(
-                    venvpath
-                ),  # put all the resulting files in that specific dir
-            ]
-            if _pip_needs_system():
-                logger.debug("adding --system to work around pip3 defaulting to --user")
-                cmd.append("--system")
-            for reqspath in self.requirement_paths:
-                cmd.append(
-                    "--requirement={}".format(reqspath)
-                )  # the dependencies file(s)
-            retcode = polite_exec(cmd)
-            if retcode:
-                raise CommandError("problems installing dependencies")
 
     def handle_package(self, bases_config: Optional[BasesConfiguration] = None):
         """Handle the final package creation."""
